eliminate a lot of redundancy between local_backend and spark_backend
