+++ hail/python/hail/backend/py4j_backend.py
@@ -1,21 +1,50 @@
-from typing import Mapping
+from typing import Mapping, Set, Tuple
 import abc
-import json
+import socket
+import socketserver
+import sys
+from threading import Thread
 
+import orjson
 import requests
 import py4j
-import py4j.java_gateway
+from py4j.java_gateway import JavaObject, JVMView
 
 import hail
 from hail.expr import construct_expr
 from hail.ir import finalize_randomness, JavaIR
 from hail.ir.renderer import CSERenderer
-from hail.utils.java import FatalError, Env
+from hail.utils.java import FatalError, Env, scala_package_object
 
 from .backend import ActionTag, Backend, fatal_error_from_java_error_triplet
+from ..hail_logging import Logger
 
 import http.client
 http.client._MAXLINE = 2 ** 20
+
+
+_installed = False
+_original = None
+
+
+def install_exception_handler():
+    global _installed
+    global _original
+    if not _installed:
+        _original = py4j.protocol.get_return_value
+        _installed = True
+        # The original `get_return_value` is not patched, it's idempotent.
+        patched = handle_java_exception(_original)
+        # only patch the one used in py4j.java_gateway (call Java API)
+        py4j.java_gateway.get_return_value = patched
+
+
+def uninstall_exception_handler():
+    global _installed
+    global _original
+    if _installed:
+        _installed = False
+        py4j.protocol.get_return_value = _original
 
 
 def handle_java_exception(f):
@@ -41,6 +70,72 @@ def handle_java_exception(f):
     return deco
 
 
+class SimpleServer(socketserver.ThreadingMixIn, socketserver.TCPServer):
+    daemon_threads = True
+    allow_reuse_address = True
+
+    def __init__(self, server_address, handler_class):
+        socketserver.TCPServer.__init__(self, server_address, handler_class)
+
+
+class LoggingTCPHandler(socketserver.StreamRequestHandler):
+    def handle(self):
+        for line in self.rfile:
+            sys.stderr.write(line.decode("ISO-8859-1"))
+
+
+class Log4jLogger(Logger):
+    def __init__(self, log_pkg):
+        self._log_pkg = log_pkg
+
+    def error(self, msg):
+        self._log_pkg.error(msg)
+
+    def warning(self, msg):
+        self._log_pkg.warn(msg)
+
+    def info(self, msg):
+        self._log_pkg.info(msg)
+
+
+def connect_logger(utils_package_object, host, port):
+    """
+    This method starts a simple server which listens on a port for a
+    client to connect and start writing messages. Whenever a message
+    is received, it is written to sys.stderr. The server is run in
+    a daemon thread from the caller, which is killed when the caller
+    thread dies.
+
+    If the socket is in use, then the server tries to listen on the
+    next port (port + 1). After 25 tries, it gives up.
+
+    :param str host: Hostname for server.
+    :param int port: Port to listen on.
+    """
+    server = None
+    tries = 0
+    max_tries = 25
+    while not server:
+        try:
+            server = SimpleServer((host, port), LoggingTCPHandler)
+        except socket.error:
+            port += 1
+            tries += 1
+
+            if tries >= max_tries:
+                sys.stderr.write(
+                    'WARNING: Could not find a free port for logger, maximum retries {} exceeded.'.format(max_tries))
+                return
+
+    t = Thread(target=server.serve_forever, args=())
+
+    # The thread should be a daemon so that it shuts down when the parent thread is killed
+    t.daemon = True
+
+    t.start()
+    utils_package_object.addSocketAppender(host, port)
+
+
 action_routes = {
     ActionTag.VALUE_TYPE: '/value/type',
     ActionTag.TABLE_TYPE: '/table/type',
@@ -55,10 +150,8 @@ action_routes = {
 
 
 class Py4JBackend(Backend):
-    _jbackend: py4j.java_gateway.JavaObject
-
     @abc.abstractmethod
-    def __init__(self):
+    def __init__(self, jvm: JVMView, jbackend: JavaObject, jhc: JavaObject):
         super(Py4JBackend, self).__init__()
         import base64
 
@@ -68,27 +161,66 @@ class Py4JBackend(Backend):
         # By default, py4j's version of this function does extra
         # work to support python 2. This eliminates that.
         py4j.protocol.decode_bytearray = decode_bytearray
+
+        self._jvm = jvm
+        self._hail_package = getattr(self._jvm, 'is').hail
+        self._utils_package_object = scala_package_object(self._hail_package.utils)
+        self._jbackend = jbackend
+        self._jhc = jhc
+
+        self._backend_server = self._hail_package.backend.BackendServer.apply(self._jbackend)
+        self._backend_server_port: int = self._backend_server.port()
+        self._backend_server.start()
         self._requests_session = requests.Session()
 
-    @abc.abstractmethod
+        self._registered_ir_function_names: Set[str] = set()
+
+        # This has to go after creating the SparkSession. Unclear why.
+        # Maybe it does its own patch?
+        install_exception_handler()
+        from hail.context import version
+
+        py_version = version()
+        jar_version = self._jhc.version()
+        if jar_version != py_version:
+            raise RuntimeError(f"Hail version mismatch between JAR and Python library\n"
+                               f"  JAR:    {jar_version}\n"
+                               f"  Python: {py_version}")
+
     def jvm(self):
-        pass
+        return self._jvm
 
-    @abc.abstractmethod
     def hail_package(self):
-        pass
+        return self._hail_package
 
-    @abc.abstractmethod
     def utils_package_object(self):
-        pass
+        return self._utils_package_object
+
+    @property
+    def logger(self):
+        if self._logger is None:
+            self._logger = Log4jLogger(self._utils_package_object)
+        return self._logger
+
+    def _rpc(self, action, payload) -> Tuple[bytes, str]:
+        data = orjson.dumps(payload)
+        path = action_routes[action]
+        port = self._backend_server_port
+        resp = self._requests_session.post(f'http://localhost:{port}{path}', data=data)
+        if resp.status_code >= 400:
+            error_json = orjson.loads(resp.content)
+            raise fatal_error_from_java_error_triplet(error_json['short'], error_json['expanded'], error_json['error_id'])
+        return resp.content, resp.headers.get('X-Hail-Timings', '')
 
     def persist_expression(self, expr):
-        assert self._jbackend
         t = expr.dtype
         return construct_expr(
             JavaIR(t, self._jbackend.executeLiteral(self._render_ir(expr._ir))),
             t
         )
+
+    def _is_registered_ir_function_name(self, name: str) -> bool:
+        return name in self._registered_ir_function_names
 
     def set_flags(self, **flags: Mapping[str, str]):
         available = self._jbackend.availableFlags()
@@ -106,7 +238,7 @@ class Py4JBackend(Backend):
         return {flag: self._jbackend.getFlag(flag) for flag in flags}
 
     def _add_reference_to_scala_backend(self, rg):
-        self._jbackend.pyAddReference(json.dumps(rg._config))
+        self._jbackend.pyAddReference(orjson.dumps(rg._config).decode('utf-8'))
 
     def _remove_reference_from_scala_backend(self, name):
         self._jbackend.pyRemoveReference(name)
@@ -153,7 +285,3 @@ class Py4JBackend(Backend):
 
     def _to_java_blockmatrix_ir(self, ir):
         return self._to_java_ir(ir, self._parse_blockmatrix_ir)
-
-    @property
-    def requires_lowering(self):
-        return True
